got window rendering in place
using the built-in ebiten debug print for now. usage of a truetype font works, but i could not for the life of me figure out how to disable anti-aliasing, so it looked like pure garbage when rendering at such a small font size. also, decided to upgrade everything ebiten to the latest master branch, so we got cool window resizing features.