refactor input control
controllers are in the actors package, the first controller being the PlayerController, which maps the input state to the players controls. This separates the logic of input handling away from the actor manager. Controllers emit events and alter actor state (though it should maybe only emit events? hmmm)

there will eventually be lots of different controllers; how and when they are called should be completely up to the actor manager. eventually, NPC movement/behavior will be governed by controllers.